refactor(card-browser): move 'rescheduleSelectedCards' to Fragment
